# openapi.yaml
openapi: 3.1.0
info:
  title: WinMyDispute API
  version: 3.1.0
  description: |
    GPT-facing API for the WinMyDispute free-preview and premium-upgrade flow.
    The free tier returns a likely reason-code match, a confidence read, and exactly one strategy tip.
    Premium output unlocks after a one-time Stripe Checkout payment tied to the user's email and can produce
    a professional dispute packet as JSON plus downloadable PDF, Word-compatible RTF, or text report output.
  contact:
    name: Daniel Neville
    email: pureplays@icloud.com
    url: https://pure-plays.com
  license:
    name: Proprietary
    url: https://pure-plays.com/terms
servers:
  - url: https://api.pure-plays.com
    description: Production custom domain
  - url: https://winmydispute-api.onrender.com
    description: Render deployment
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/intake/normalize:
    post:
      tags: [intake]
      summary: Normalize messy intake into canonical API fields
      description: |
        Helps the GPT turn inconsistent customer answers into the API's canonical
        structure. The endpoint normalizes flexible date and amount formats, infers
        merchant vertical when possible, and returns missing fields plus a likely
        reason-code match for follow-up questions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "200":
          description: Normalized intake and follow-up guidance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormalizedIntakeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/issuers/{issuer}/profile:
    get:
      tags: [issuers]
      summary: Return richer issuer filing guidance and profile data
      parameters:
        - name: issuer
          in: path
          required: true
          schema:
            type: string
          example: Chase
      responses:
        "200":
          description: Issuer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Issuer profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/reasons/search:
    get:
      tags: [reasons]
      summary: Search reason-code profiles by natural language query
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          example: canceled subscription renewed anyway
        - name: network
          in: query
          required: false
          schema:
            type: string
            enum: [visa, mastercard, amex, discover]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 8
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasonSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/cases:
    get:
      tags: [disputes]
      summary: List saved premium case files for an email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: premiumAccessToken
          in: query
          required: false
          schema:
            type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
          description: Matching Stripe Checkout session ID for token exchange.
        - name: checkoutSessionId
          in: query
          required: false
          schema:
            type: string
          description: Alias for sessionId.
      responses:
        "200":
          description: Saved case list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCasesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before case access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/cases/{caseId}:
    get:
      tags: [disputes]
      summary: Retrieve a saved premium case file
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: premiumAccessToken
          in: query
          required: false
          schema:
            type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
          description: Matching Stripe Checkout session ID for token exchange.
        - name: checkoutSessionId
          in: query
          required: false
          schema:
            type: string
          description: Alias for sessionId.
      responses:
        "200":
          description: Saved case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFileDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before case access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "404":
          description: Case file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/evidence/quality-score:
    post:
      tags: [disputes]
      summary: Score how submission-ready the current evidence package is
      description: |
        Gives the GPT a structured readiness check before premium generation or filing.
        The endpoint scores evidence completeness, compares provided material to the
        strongest expected evidence, and returns the highest-priority missing items.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DisputeInput"
                - type: object
                  properties:
                    reasonCode:
                      type: string
      responses:
        "200":
          description: Evidence-quality summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceQualityScoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/disputes/estimate-success:
    post:
      tags: [disputes]
      summary: Estimate the likelihood of dispute success
      description: |
        Returns a heuristic estimate by default and blends in anonymized historical outcome feedback
        when there is enough matching data for the same reason code, issuer, or merchant vertical.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateSuccessRequest"
      responses:
        "200":
          description: Success estimate response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/disputes/outcome-feedback:
    post:
      tags: [analytics]
      summary: Record anonymized dispute outcome feedback
      description: |
        Stores outcome data that can later improve success-estimate quality for similar disputes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutcomeFeedbackRequest"
      responses:
        "201":
          description: Outcome feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeFeedbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Request limit exceeded for the current rate-limit window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: A valid email address is required.
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [error]
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: Too many requests
        requestId:
          type: string
          format: uuid
          nullable: true
        retryAfterSeconds:
          type: integer
          minimum: 1
          example: 42
      required: [error, requestId, retryAfterSeconds]
    EvidenceItem:
      oneOf:
        - type: string
        - type: object
          additionalProperties: true
          properties:
            summary:
              type: string
            description:
              type: string
            text:
              type: string
            filename:
              type: string
            fileType:
              type: string
            type:
              type: string
            extractedText:
              type: string
    DisputeInput:
      type: object
      additionalProperties: true
      properties:
        email:
          type: string
          format: email
          description: Optional for preview, required for premium generation and license lookup.
        premiumAccessToken:
          type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        checkoutSessionId:
          type: string
          description: Matching Stripe Checkout session ID that can be exchanged for a premium access token.
        sessionId:
          type: string
          description: Alias for checkoutSessionId.
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        issuer:
          type: string
          example: Chase
        cardholderName:
          type: string
          example: Jane Doe
        addressLine1:
          type: string
          example: 123 Main St
        addressLine2:
          type: string
          example: Miami, FL 33101
        phone:
          type: string
          example: "3055551212"
        merchantName:
          type: string
          example: Example Merchant
        merchantVertical:
          type: string
          example: subscription
        transactionAmount:
          oneOf:
            - type: number
            - type: string
          example: $1,234.56
          description: Flexible amount input. The API normalizes common U.S. and international display formats.
        transactionDate:
          type: string
          example: 03/01/2026
          description: Flexible date input. The API normalizes common slash, dash, dot, and month-name formats.
        description:
          type: string
          description: Primary dispute summary used for reason-code matching.
          example: I cancelled before renewal but the merchant chargd me anyway.
        evidenceSummary:
          type: string
          example: I have cancellation emails and screenshots of the account page.
        evidenceItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        timelineItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        merchantRebuttalConcerns:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        desiredOutcome:
          type: string
          example: Reverse the charge and confirm the billing correction in writing.
        tone:
          type: string
          enum: [formal, assertive, concise, empathetic]
          example: formal
        lengthPreference:
          type: string
          enum: [short, standard, detailed]
          example: detailed
        outputFormat:
          type: string
          enum: [pdf, docx, word-compatible-rtf, text]
          example: pdf
        async:
          type: boolean
          description: When true on supported premium endpoints, the server returns a queued job instead of waiting for completion.
          example: true
        redactionMode:
          type: string
          enum: [none, standard, strict]
          description: Optional privacy mode for generated report artifacts.
          example: standard
        includeRedactedVersion:
          type: boolean
          description: When true, the report-document endpoint returns both a full copy and a redacted copy.
          example: true
      required: [description]
    NormalizedIntake:
      type: object
      properties:
        caseId:
          type: string
        email:
          type: string
          format: email
          nullable: true
        network:
          type: string
          nullable: true
        issuer:
          type: string
        merchantName:
          type: string
        merchantVertical:
          type: string
          nullable: true
        transactionAmount:
          type: string
        transactionAmountValue:
          type: number
          nullable: true
        transactionDate:
          type: string
        transactionDateIso:
          type: string
          nullable: true
        description:
          type: string
        evidenceSummary:
          type: string
        evidenceItems:
          type: array
          items:
            type: string
        timelineItems:
          type: array
          items:
            type: string
        merchantRebuttalConcerns:
          type: array
          items:
            type: string
        desiredOutcome:
          type: string
        tone:
          type: string
        lengthPreference:
          type: string
        outputFormat:
          type: string
        redactionMode:
          type: string
        includeRedactedVersion:
          type: boolean
      required:
        - caseId
        - network
        - issuer
        - merchantName
        - transactionAmount
        - transactionDate
        - description
        - evidenceItems
        - timelineItems
        - merchantRebuttalConcerns
        - desiredOutcome
        - tone
        - lengthPreference
        - outputFormat
        - redactionMode
        - includeRedactedVersion
    NormalizedIntakeResponse:
      type: object
      properties:
        normalizedIntake:
          $ref: "#/components/schemas/NormalizedIntake"
        missingFields:
          type: array
          items:
            type: string
        normalizationNotes:
          type: array
          items:
            type: string
        reasonMatch:
          $ref: "#/components/schemas/ReasonMatchSummary"
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [normalizedIntake, missingFields, normalizationNotes, reasonMatch]
    ReasonMatchSummary:
      type: object
      properties:
        reasonCode:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        network:
          type: string
          nullable: true
        confidence:
          type: integer
      required: [reasonCode, title, description, network, confidence]
    ReasonSearchResult:
      type: object
      properties:
        network:
          type: string
        reasonCode:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        strategyTips:
          type: array
          items:
            type: string
        evidenceRequirements:
          type: array
          items:
            type: string
        score:
          type: integer
      required: [network, reasonCode, title, strategyTips, evidenceRequirements, score]
    ReasonSearchResponse:
      type: object
      properties:
        query:
          type: string
        network:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/ReasonSearchResult"
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [query, network, results]
...
    IssuerProfile:
      type: object
      properties:
        issuer:
          type: string
        aliases:
          type: array
          items:
            type: string
        supportedNetworks:
          type: array
          items:
            type: string
        contact:
          $ref: "#/components/schemas/IssuerGuidance"
        notes:
          type: array
          items:
            type: string
        evidenceExamples:
          type: array
          items:
            type: string
        preferredSubmissionChannels:
          type: array
          items:
            type: string
      required:
        - issuer
        - aliases
        - supportedNetworks
        - contact
        - notes
        - evidenceExamples
        - preferredSubmissionChannels
...
    EvidenceQualityScoreResponse:
      type: object
      properties:
        reasonCode:
          type: string
          nullable: true
        reasonTitle:
          type: string
          nullable: true
        merchantVertical:
          type: string
          nullable: true
        evidenceQualityScore:
          type: integer
          minimum: 0
          maximum: 100
        readinessLevel:
          type: string
          enum: [strong, good, needs-more-evidence, thin]
        providedEvidenceCount:
          type: integer
        timelineItemCount:
          type: integer
        recommendedEvidence:
          type: array
          items:
            type: string
        coveredRecommendations:
          type: array
          items:
            type: string
        missingPriorityEvidence:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        requestId:
          type: string
          format: uuid
          nullable: true
      required:
        - reasonCode
        - reasonTitle
        - merchantVertical
        - evidenceQualityScore
        - readinessLevel
        - providedEvidenceCount
        - timelineItemCount
        - recommendedEvidence
        - coveredRecommendations
        - missingPriorityEvidence
        - warnings
...
    ListCasesResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        count:
          type: integer
        cases:
          type: array
          items:
            $ref: "#/components/schemas/CaseFileSummary"
      required: [email, count, cases]
    CaseFileDetail:
      allOf:
        - $ref: "#/components/schemas/CaseFileSummary"
        - type: object
          properties:
            premiumAccessToken:
              type: string
              nullable: true
            versions:
              type: array
              items:
                $ref: "#/components/schemas/CaseVersionSummary"
          required: [versions]
...
    EstimateSuccessRequest:
      type: object
      properties:
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        reasonCode:
          type: string
        issuer:
          type: string
        merchantVertical:
          type: string
        consumerEvidence:
          type: string
        priorAttemptsToResolve:
          type: boolean
        merchantResponse:
          type: string
        transactionAmount:
          oneOf:
            - type: number
            - type: string
      required: [network, reasonCode]
    EstimateSuccessResponse:
      type: object
      properties:
        estimatedSuccessRate:
          type: number
          minimum: 0
          maximum: 1
          example: 0.73
        rationale:
          type: string
        modelType:
          type: string
          enum: [heuristic, blended-historical-and-heuristic]
        sampleSize:
          type: integer
          example: 5
      required: [estimatedSuccessRate, rationale, modelType, sampleSize]
    OutcomeFeedbackRequest:
      type: object
      properties:
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        reasonCode:
          type: string
        issuer:
          type: string
        merchantVertical:
          type: string
        merchantType:
          type: string
        outcome:
          type: string
          enum: [won, lost, partial, reversed, denied]
        amount:
          oneOf:
            - type: number
            - type: string
        notes:
          type: string
      required: [network, reasonCode, outcome]
    OutcomeFeedbackResponse:
      type: object
      properties:
        recorded:
          type: boolean
          example: true
        outcome:
          type: string
          example: won
        reasonCode:
          type: string
          example: 13.2
        network:
          type: string
          example: visa
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [recorded, outcome, reasonCode, network, requestId]
    UpgradeRequiredResponse:
      type: object
      properties:
        upgradeRequired:
          type: boolean
          example: true
        checkoutUrl:
          type: string
          format: uri
          example: https://checkout.stripe.com/pay/cs_test_123
        checkoutSessionId:
          type: string
          example: cs_test_123
        message:
          type: string
          example: Premium access requires a one-time $6.99 payment.
      required: [upgradeRequired, checkoutUrl, checkoutSessionId, message]